Copy /proc/[pid]/exe from parent to child at fork()
