+++ include/myst/procfs.h
@@ -13,6 +13,9 @@ int create_proc_root_entries();
 /* For callbacks implementing /proc/[pid]/xxx entries */
 myst_process_t* myst_procfs_path_to_process(const char* entrypath);
 
+/* Create the "/proc/<pid>/exe" link */
+int procfs_setup_exe_link(const char* path, pid_t pid);
+
 /*
 **==============================================================================
 **
